# docs/swagger.yaml
definitions:
  errs.Action:
    properties:
      message:
        type: string
      type:
        $ref: '#/definitions/errs.ActionType'
      value:
        type: string
    type: object
  errs.ActionType:
    enum:
    - redirect
    type: string
    x-enum-varnames:
    - ActionTypeRedirect
  errs.FieldError:
    properties:
      error:
        type: string
      field:
        type: string
    type: object
  errs.HTTPError:
    properties:
      action:
        allOf:
        - $ref: '#/definitions/errs.Action'
        description: action to be taken
      code:
        type: string
      errors:
        description: field level errors
        items:
          $ref: '#/definitions/errs.FieldError'
        type: array
      message:
        type: string
      override:
        type: boolean
      status:
        type: integer
    type: object
...
  model.AnswerResponse:
    properties:
      accepted_answers:
        items:
          type: string
        type: array
      case_sensitive:
        type: boolean
      correct_bool:
        type: boolean
      correct_choice_id:
        type: string
      correct_choice_position:
        type: integer
      explanation:
        type: string
      model_answer:
        type: string
    type: object
...
  model.ChoiceResponse:
    properties:
      choice_id:
        type: string
      position:
        type: integer
      text:
        type: string
    type: object
...
  model.Exam:
    properties:
      camera_required:
        type: boolean
      created_at:
        type: string
      created_by:
        $ref: '#/definitions/users.User'
      description:
        type: string
      duration:
        type: integer
      end_time:
        type: string
      id:
        type: string
      invite_code:
        type: string
      max_violations:
        type: integer
      proctoring_enabled:
        type: boolean
      session_status:
        type: string
      set:
        $ref: '#/definitions/model.Set'
      set_id:
        type: string
      start_mode:
        type: string
      start_time:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      visibility:
        type: string
    type: object
...
  model.QuestionResponse:
    properties:
      answer:
        $ref: '#/definitions/model.AnswerResponse'
      answer_id:
        type: string
      choices:
        items:
          $ref: '#/definitions/model.ChoiceResponse'
        type: array
      difficulty:
        type: string
      question_id:
        type: string
      text:
        type: string
      type:
        type: string
    type: object
  model.Set:
    properties:
      created_at:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
      visibility:
        type: string
    type: object
...
  users.User:
    properties:
      email:
        type: string
      id:
        type: string
      image_url:
        type: string
      name:
        type: string
    type: object
info:
  contact: {}
paths:
  /api/v1/exams/:
    get:
      description: Returns exams filtered by set_id (if provided) or all exams for
        the authenticated user
      parameters:
      - description: Optional set ID to filter exams
        in: query
        name: set_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Exam'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errs.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: List exams
      tags:
      - Exams
    post:
      consumes:
      - application/json
      description: Validates input and creates a new exam from a question set
      parameters:
      - description: Exam creation payload (start_time in RFC3339 format)
        in: body
        name: body
        required: true
        schema:
          properties:
            camera_required:
              type: boolean
            description:
              type: string
            duration_in_minutes:
              type: integer
            invite_code:
              type: string
            proctoring_enabled:
              type: boolean
            session_status:
              type: string
            set_id:
              type: string
            start_mode:
              type: string
            start_time:
              type: string
            title:
              type: string
          type: object
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errs.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: Create an exam
      tags:
      - Exams
  /api/v1/exams/{exam_id}:
    delete:
      description: Deletes an exam by its ID
      parameters:
      - description: Exam ID
        in: path
        name: exam_id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errs.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: Delete an exam
      tags:
      - Exams
    get:
      description: Returns a single exam by its ID
      parameters:
      - description: Exam ID
        in: path
        name: exam_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Exam'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errs.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: Get an exam by ID
      tags:
      - Exams
  /api/v1/exams/q/{exam_id}:
    get:
      description: Returns all questions with choices and answers for the specified
        exam
      parameters:
      - description: Exam ID
        in: path
        name: exam_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.QuestionResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/errs.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/errs.HTTPError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/errs.HTTPError'
      summary: Get questions of an exam
      tags:
      - Exams
...
swagger: "2.0"